+++ CHANGELOG.md
@@ -336,3 +336,4 @@
 - [2026-04-12 11:38] feat(i18n): implement responsive i18n view component
 - [2026-04-12 12:28] feat(ui): implement shimmer loading skeleton in ui
 - [2026-04-12 12:51] test(storage): add unit tests for storage business logic
+- [2026-04-12 13:17] style(cache): harmonize light and dark theme palette for cache